+++ takeout_lib/lib/browser/provider.dart
@@ -20,6 +20,7 @@ import 'dart:convert';
 import 'package:audio_service/audio_service.dart';
 import 'package:crypto/crypto.dart';
 import 'package:takeout_lib/api/model.dart';
+import 'package:takeout_lib/cache/offset_repository.dart';
 import 'package:takeout_lib/cache/spiff.dart';
 import 'package:takeout_lib/client/repository.dart';
 import 'package:takeout_lib/history/model.dart';
@@ -29,6 +30,83 @@ import 'package:takeout_lib/media_type/repository.dart';
 import 'package:takeout_lib/settings/repository.dart';
 import 'package:takeout_lib/spiff/model.dart';
 import 'package:takeout_lib/subscribed/repository.dart';
+
+/*
+
+This table shows the media hierarchy, managed as a file system
+with paths that returns lists of MediaItems that are playable
+and non-playable. Podcasts items include progress details.
+
+Support for movie playback is TBD. Not sure right now if it's even possible.
+
++---------+---------------------------+--------+
+|Recent   |/history                   |List    |
++---------+---------------------------+--------+
+|         |/history/spiffs/{id}       |Playable|
++---------+---------------------------+--------+
+|Music    |/music                     |List    |
++---------+---------------------------+--------+
+|         |/music/releases/{id}/tracks|Playable|
++---------+---------------------------+--------+
+|         |/music/radio/stations/{id} |Playable|
++---------+---------------------------+--------+
+|Podcasts |/podcasts                  |Grid    |
++---------+---------------------------+--------+
+|         |/podcasts/series/{id}      |Grid    |
++---------+---------------------------+--------+
+|         |/podcasts/episodes/{id}    |Playable|
++---------+---------------------------+--------+
+|Radio    |/radio                     |List    |
++---------+---------------------------+--------+
+|         |/radio/genre               |List    |
++---------+---------------------------+--------+
+|         |/radio/period              |List    |
++---------+---------------------------+--------+
+|         |/radio/other               |List    |
++---------+---------------------------+--------+
+|         |/radio/stream              |List    |
++---------+---------------------------+--------+
+|Artists  |/artists                   |List    |
++---------+---------------------------+--------+
+|         |/artists/{id}              |Grid    |
++---------+---------------------------+--------+
+|Playlists|/playlists                 |List    |
++---------+---------------------------+--------+
+|         |/playlists/{id}            |Playable|
++---------+---------------------------+--------+
+|Downloads|/downloads                 |List    |
++---------+---------------------------+--------+
+|         |/downloads/spiffs/{id}     |Playable|
++---------+---------------------------+--------+
+|Movies   |/movies                    |Grid    |
++---------+---------------------------+--------+
+|         |/movies/{id}               |Playable|
++---------+---------------------------+--------+
+
+ */
+
+// TYPE: double, a value between 0.0 and 1.0, inclusive. 0.0 indicates no
+// completion progress (item is not started) and 1.0 indicates full completion
+// progress (item is fully played). Values in between indicate partial progress
+// (for example, 0.75 indicates the item is 75% complete).
+const extrasKeyCompletionPercentage =
+    'androidx.media.MediaItem.Extras.COMPLETION_PERCENTAGE';
+
+const extrasKeyCompletionStatus = 'android.media.extra.PLAYBACK_STATUS';
+const extrasValueCompletionStatusNotPlayed = 0;
+const extrasValueCompletionStatusPartialPlayed = 1;
+const extrasValueCompletionStatusFullyPlayed = 2;
+
+const extrasKeyContentBrowsableStyle =
+    'android.media.browse.CONTENT_STYLE_BROWSABLE_HINT';
+const extrasKeyContentPlayableStyle =
+    'android.media.browse.CONTENT_STYLE_PLAYABLE_HINT';
+const extrasKeyContentSingleItemStyle =
+    'android.media.browse.CONTENT_STYLE_SINGLE_ITEM_HINT';
+const extrasValueContentStyleListItem = 1;
+const extrasValueContentStyleGridItem = 2;
+const extrasValueContentStyleCategoryListItem = 3;
+const extrasValueContentStyleCategoryGridItem = 4;
 
 abstract class MediaProvider {
   Future<List<MediaItem>> getRoot();
@@ -53,6 +131,7 @@ class DefaultMediaProvider implements MediaProvider {
   final SpiffCacheRepository spiffCacheRepository;
   final MediaTypeRepository mediaTypeRepository;
   final SubscribedRepository subscribedRepository;
+  final OffsetCacheRepository offsetCacheRepository;
 
   DefaultMediaProvider(
       this.clientRepository,
@@ -60,7 +139,8 @@ class DefaultMediaProvider implements MediaProvider {
       this.settingsRepository,
       this.spiffCacheRepository,
       this.mediaTypeRepository,
-      this.subscribedRepository);
+      this.subscribedRepository,
+      this.offsetCacheRepository);
 
   @override
   Future<List<MediaItem>> getRoot() async {
@@ -68,21 +148,66 @@ class DefaultMediaProvider implements MediaProvider {
     final index = await clientRepository.index();
     // TODO need to localize these strings. Will need a repository for that
     // since there's no context available here.
-    items.add(const MediaItem(id: '/history', title: 'Recent', playable: false));
+    items.add(const MediaItem(
+      id: '/history',
+      title: 'Recent',
+      playable: false,
+    ));
     if (index.hasMusic) {
-      items.add(const MediaItem(id: '/music', title: 'Music', playable: false));
+      items.add(const MediaItem(
+        id: '/music',
+        title: 'Music',
+        playable: false,
+      ));
     }
     if (index.hasPodcasts) {
-      items.add(const MediaItem(id: '/podcasts', title: 'Podcasts', playable: false));
+      items.add(const MediaItem(
+        id: '/podcasts',
+        title: 'Podcasts',
+        playable: false,
+        extras: {
+          extrasKeyContentBrowsableStyle: extrasValueContentStyleGridItem,
+          extrasKeyContentPlayableStyle: extrasValueContentStyleGridItem,
+        },
+      ));
     }
     if (index.hasMusic) {
-      items.add(const MediaItem(id: '/radio', title: 'Radio', playable: false));
-      items.add(const MediaItem(id: '/artists', title: 'Artists', playable: false));
+      items.add(const MediaItem(
+        id: '/radio',
+        title: 'Radio',
+        playable: false,
+      ));
+      items.add(const MediaItem(
+        id: '/artists',
+        title: 'Artists',
+        playable: false,
+      ));
+    }
+    if (index.hasPlaylists) {
+      items.add(const MediaItem(
+        id: '/playlists',
+        title: 'Playlists',
+        playable: false,
+      ));
     }
     final downloads = await spiffCacheRepository.entries;
-    if (downloads.isNotEmpty && downloads.any((d) => d.isMusic())) {
-      items.add(
-          const MediaItem(id: '/downloads', title: 'Downloads', playable: false));
+    if (downloads.isNotEmpty) {
+      items.add(const MediaItem(
+        id: '/downloads',
+        title: 'Downloads',
+        playable: false,
+      ));
+    }
+    if (index.hasMovies) {
+      items.add(const MediaItem(
+        id: '/movies',
+        title: 'Movies',
+        playable: false,
+        extras: {
+          extrasKeyContentBrowsableStyle: extrasValueContentStyleGridItem,
+          extrasKeyContentPlayableStyle: extrasValueContentStyleGridItem,
+        },
+      ));
     }
     return items;
   }
@@ -107,8 +232,12 @@ class DefaultMediaProvider implements MediaProvider {
         return _getRadio();
       case '/artists':
         return _getArtists();
+      case '/playlists':
+        return _getPlaylists();
       case '/downloads':
         return _getDownloads();
+      case '/movies':
+        return _getMovies();
       default:
         if (parentId.startsWith('/artists/')) {
           return _getArtist(parentId);
@@ -128,10 +257,10 @@ class DefaultMediaProvider implements MediaProvider {
 
     if (mediaType == MediaType.video) {
       for (var m in results.movies ?? <Movie>[]) {
-        items.add(_movie(m));
+        items.add(await _movie(m));
       }
     } else if (mediaType == MediaType.podcast) {
-      for (var s in results.series ??<Series> []) {
+      for (var s in results.series ?? <Series>[]) {
         items.add(_series(s));
       }
     } else if (mediaType == MediaType.stream) {
@@ -189,6 +318,9 @@ class DefaultMediaProvider implements MediaProvider {
       spiff = await _getHistorySpiff(mediaId);
     } else if (mediaId.startsWith('/downloads/spiffs/')) {
       spiff = await _getDownloadSpiff(mediaId);
+    } else if (mediaId.startsWith('/playlists/')) {
+      final id = int.parse(mediaId.split('/')[2]);
+      spiff = await clientRepository.playlist(id: id);
     }
     return spiff;
   }
@@ -219,16 +351,22 @@ class DefaultMediaProvider implements MediaProvider {
     return null;
   }
 
+  Future<List<MediaItem>> _getPlaylists() async {
+    final items = <MediaItem>[];
+    final view = await clientRepository.playlists();
+    for (var p in view.playlists) {
+      items.add(await _playlist(p));
+    }
+    return items;
+  }
+
   Future<List<MediaItem>> _getDownloads() async {
     final items = <MediaItem>[];
     final downloads = await spiffCacheRepository.entries;
     final entries = List<Spiff>.from(downloads);
     entries.sort((a, b) => a.title.compareTo(b.title));
     for (var d in entries) {
-      if (d.isMusic()) {
-        // FIXME download keys appear to be the same
-        items.add(_download(d));
-      }
+      items.add(await _download(d));
     }
     return items;
   }
@@ -239,7 +377,7 @@ class DefaultMediaProvider implements MediaProvider {
     final spiffs = List<SpiffHistory>.from(history.spiffs);
     spiffs.sort((a, b) => b.dateTime.compareTo(a.dateTime));
     for (var s in spiffs) {
-      items.add(_history(s));
+      items.add(await _history(s));
     }
     return items;
   }
@@ -277,19 +415,20 @@ class DefaultMediaProvider implements MediaProvider {
     final items = <MediaItem>[];
     final radio = await clientRepository.radio();
     if (radio.genre != null) {
-      items
-          .add(const MediaItem(id: '/radio/genre', title: 'Genres', playable: false));
+      items.add(const MediaItem(
+          id: '/radio/genre', title: 'Genres', playable: false));
     }
     if (radio.period != null) {
-      items.add(
-          const MediaItem(id: '/radio/period', title: 'Decades', playable: false));
+      items.add(const MediaItem(
+          id: '/radio/period', title: 'Decades', playable: false));
     }
     if (radio.other != null) {
-      items.add(const MediaItem(id: '/radio/other', title: 'Other', playable: false));
+      items.add(
+          const MediaItem(id: '/radio/other', title: 'Other', playable: false));
     }
     if (radio.stream != null) {
-      items.add(
-          const MediaItem(id: '/radio/stream', title: 'Streams', playable: false));
+      items.add(const MediaItem(
+          id: '/radio/stream', title: 'Streams', playable: false));
     }
     return items;
   }
@@ -312,6 +451,36 @@ class DefaultMediaProvider implements MediaProvider {
         items.add(_station(s));
       }
     }
+    return items;
+  }
+
+  Future<List<MediaItem>> _getMovies() async {
+    final items = <MediaItem>[];
+    final home = await clientRepository.home();
+
+    Iterable<Movie>? movies;
+    final videoType = mediaTypeRepository.videoType;
+    switch (videoType) {
+      case VideoType.added:
+        movies = home.addedMovies;
+      case VideoType.recent:
+        movies = home.newMovies;
+      case VideoType.recommended:
+        final recommended = home.recommendMovies;
+        if (recommended != null) {
+          movies = recommended.first.movies ?? [];
+        }
+      case VideoType.all:
+        final view = await clientRepository.movies();
+        movies = List<Movie>.from(view.movies);
+    }
+
+    if (movies != null) {
+      for (var m in movies) {
+        items.add(await _movie(m));
+      }
+    }
+
     return items;
   }
 
@@ -342,7 +511,7 @@ class DefaultMediaProvider implements MediaProvider {
     final id = int.parse(parentId.split('/')[3]);
     final series = await clientRepository.series(id);
     for (var e in series.episodes) {
-      items.add(_episode(e));
+      items.add(await _episode(e));
     }
     return items;
   }
@@ -394,25 +563,71 @@ class DefaultMediaProvider implements MediaProvider {
         playable: false);
   }
 
-  MediaItem _episode(Episode e) {
+  Future<Map<String, dynamic>> _completionExtras(OffsetIdentifier id) async {
+    final offset = await offsetCacheRepository.get(id);
+    final completionPercentage = offset?.value();
+    var completionStatus = extrasValueCompletionStatusNotPlayed;
+
+    if (completionPercentage != null) {
+      if (completionPercentage >= 95) {
+        completionStatus = extrasValueCompletionStatusFullyPlayed;
+      } else if (completionPercentage > 0) {
+        completionStatus = extrasValueCompletionStatusPartialPlayed;
+      }
+    }
+
+    return {
+      if (completionPercentage != null)
+        extrasKeyCompletionPercentage: completionPercentage,
+      extrasKeyCompletionStatus: completionStatus,
+    };
+  }
+
+  Future<MediaItem> _episode(Episode e) async {
+    final extras = await _completionExtras(e);
+    // no effect?
+    // extras[extrasKeyContentPlayableStyle] = extrasValueContentStyleGridItem;
     return MediaItem(
-        id: '/podcasts/episodes/${e.id}',
-        title: e.title,
-        artist: e.author,
-        artUri: _img(e.image),
-        playable: true);
+      id: '/podcasts/episodes/${e.id}',
+      title: e.title,
+      artist: e.author,
+      artUri: _img(e.image),
+      playable: true,
+      extras: extras,
+    );
   }
 
   MediaItem _artist(Artist a) {
-    return MediaItem(id: '/artists/${a.id}', title: a.name, playable: false);
+    return MediaItem(
+      id: '/artists/${a.id}',
+      title: a.name,
+      playable: false,
+      extras: {
+        // use grid for artist releases
+        extrasKeyContentBrowsableStyle: extrasValueContentStyleGridItem,
+        extrasKeyContentPlayableStyle: extrasValueContentStyleGridItem,
+      },
+    );
   }
 
-  MediaItem _history(SpiffHistory h) {
+  Future<Map<String, dynamic>?> _spiffExtras(Spiff spiff) async {
+    Map<String, dynamic>? extras;
+
+    if (spiff.isPodcast() && spiff.length == 1) {
+      final episode = spiff[0];
+      extras = await _completionExtras(episode);
+    }
+
+    return extras;
+  }
+
+  Future<MediaItem> _history(SpiffHistory h) async {
     return MediaItem(
         id: '/history/spiffs/${h.dateTime.millisecondsSinceEpoch}',
         title: h.spiff.title,
         artUri: _img(h.spiff.cover),
-        playable: true);
+        playable: true,
+        extras: await _spiffExtras(h.spiff));
   }
 
   MediaItem _track(Track t) {
@@ -423,12 +638,14 @@ class DefaultMediaProvider implements MediaProvider {
         playable: true);
   }
 
-  MediaItem _movie(Movie m) {
+  Future<MediaItem> _movie(Movie m) async {
     return MediaItem(
-        id: '/movies/${m.id}',
-        title: m.title,
-        artUri: _img(m.image),
-        playable: true);
+      id: '/movies/${m.id}',
+      title: m.title,
+      artUri: _img(m.image),
+      playable: true,
+      extras: await _completionExtras(m),
+    );
   }
 
   String _md5(String input) {
@@ -439,12 +656,22 @@ class DefaultMediaProvider implements MediaProvider {
     return _md5('${spiff.location}${spiff.date}');
   }
 
-  MediaItem _download(Spiff spiff) {
+  Future<MediaItem> _download(Spiff spiff) async {
     return MediaItem(
-        id: '/downloads/spiffs/${_downloadKey(spiff)}',
-        title: spiff.title,
-        artUri: _img(spiff.cover),
-        playable: true);
+      id: '/downloads/spiffs/${_downloadKey(spiff)}',
+      title: spiff.title,
+      artUri: _img(spiff.cover),
+      playable: true,
+      extras: await _spiffExtras(spiff),
+    );
+  }
+
+  Future<MediaItem> _playlist(PlaylistView playlist) async {
+    return MediaItem(
+      id: '/playlists/${playlist.id}',
+      title: playlist.name,
+      playable: true,
+    );
   }
 
   Uri _img(String i) {
@@ -456,10 +683,6 @@ class DefaultMediaProvider implements MediaProvider {
   }
 
   Future<SearchView> _search(String query) async {
-    if (query.contains(RegExp(r'[:"\\*]')) == false) {
-      query =
-          'title:"$query*" release:"$query*" artist:"$query*" genre:"$query*"';
-    }
     return await clientRepository.search(query);
   }
 }